% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_method_ensemble_function.R
\name{age_ensemble}
\alias{age_ensemble}
\title{Predict age using Twitter profile data}
\usage{
age_ensemble(input, labeled = FALSE, set.seed = NULL,
  training_set = 0.8, chunk = FALSE, action = "append", dir = NULL)
}
\arguments{
\item{input}{A data frame containing Twitter profile information. Created using the function get_user_info.}

\item{chunk}{If TRUE, features and predictions are generated for 10,000 user chunks. Recommended for unlabeled data with >10,000 users.}

\item{action}{If chunk==TRUE, whether the data should be appended and returned as one file, or saved as chunks. Options are append and save.}

\item{dir}{Where should output files be saved? Used only if action==save. Defaults to NULL}

\item{ncores}{Number of cores across which the process is distributed. Note: Windows users cannot request more than two.}
}
\value{
results A two-column data frame containing Twitter user IDs and age predictions.
}
\description{
This function inputs Twitter profile metadata from get_user_info, generates features from the profile content and last posted status, and uses a gradient boosted decision tree to identify whether users are Below 18, 18-30 or 30 and older.
}
\examples{
## df_list from function get_user_info
## df <- do.call("rbind", df_list)
gender_ensemble(df, labeled=FALSE, set.seed=NULL, training_set=0.80, chunk=TRUE, action="save", dir="/home/User/workingDir")
}
\keyword{demography}
\keyword{gender,}
\keyword{twitter,}
