% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonant_count.R
\name{consonant_count}
\alias{consonant_count}
\title{Count consonants in a string}
\usage{
consonant_count(name)
}
\arguments{
\item{name}{A string value}
}
\value{
results A numeric value representing the number of consonants
}
\description{
Count the number of consonants in a string
}
\examples{
consonant_count("Jim")
}
\keyword{language}
\keyword{text,}
