% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_names.R
\name{convert_names}
\alias{convert_names}
\title{Clean user names}
\usage{
convert_names(name)
}
\arguments{
\item{vect}{A character vector containing a Twitter user name}
}
\value{
name_new a cleaned first name from a user name
}
\description{
Clean and extract first names from user names
}
\examples{
convert_names("Miss. Jessica is the greatest!")
}
\keyword{Twitter}
\keyword{text,}
