% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corp_convert.R
\name{corp_convert}
\alias{corp_convert}
\title{Prepare a corpus of tweets}
\usage{
corp_convert(vector)
}
\arguments{
\item{vector}{A vector of tweets}
}
\value{
corp A clean corpus of tweet data, ready to be analyzed and converted to a document-term or term-document matrix.
}
\description{
Prepare a vector of tweets for conversion to a document-term matrix or term-document matrix. First convert to ASCII text. Then stem, normalize, remove stopwords, remove punctuation, remove URLs, remove numbers and strip whitespace
}
\keyword{Twitter}
\keyword{text,}
