% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpaces.R
\name{countSpaces}
\alias{countSpaces}
\title{Count spaces in a string}
\usage{
countSpaces(s)
}
\arguments{
\item{s}{A string value or vector or strings}
}
\value{
results A numeric value representing the number of spaces
}
\description{
Count the number of spaces in a string
}
\examples{
countSpaces("Check how many spaces are in this sentence.")
}
\seealso{
This solution first posted on StackOverflow by user MvG
}
\keyword{language}
\keyword{text,}
