% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user_signature.R
\name{create_user_signature}
\alias{create_user_signature}
\title{Authenticate connection to the Twitter API: Part II}
\usage{
create_user_signature(api_key, api_secret, acc_token, acc_token_secret,
  save = FALSE, dir = NULL)
}
\arguments{
\item{api_key}{The Consumer Key value associated with your Twitter app}

\item{api_secret}{Consumer secret associated with your Twitter app}

\item{acc_token}{Access token associated with your Twitter app}

\item{acc_token_secret}{Access token secret associated with your Twitter app}

\item{save}{A TRUE/FALSE indication of whether you want to auto-save your signature as an .RData file for later use}

\item{acc_token_secret}{If save==TRUE, your preferred directory}
}
\value{
user.signature A user signature that you can use and save
}
\description{
This function allows users to authenticate a connection to the Twitter API. To establish an API connection, you will need a Twitter Developer's Account and an application. See https://apps.twitter.com/ for details.
}
\keyword{twitter}
