% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateConvert.R
\name{dateConvert}
\alias{dateConvert}
\title{Convert date formats}
\usage{
dateConvert(date)
}
\arguments{
\item{x}{a string value with a time stamp formatted "%a %b %d %h:%m:%s %z %Y"}
}
\value{
newDate A Date object formatted "%Y-%M-%D"
}
\description{
Convert the created_at timestamp on a tweet into a usable Date object
}
\examples{
dateConvert("Fri Jan 20 21:54:23 +0000 2012")
}
\keyword{TWitter}
\keyword{date,}
