% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_char.R
\name{end_char}
\alias{end_char}
\title{Vowel or consonant as last character}
\usage{
end_char(name)
}
\arguments{
\item{input}{A character value}
}
\value{
A value of "consonant" or "vowel"
}
\description{
This function evaluates characters and termines whether the last term is a vowel or consonant.
}
\examples{
end_char("jim")
}
\keyword{character,}
\keyword{language}
