% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english_syllable_count.R
\name{english_syllable_count}
\alias{english_syllable_count}
\title{Count syllables}
\usage{
english_syllable_count(ortho)
}
\arguments{
\item{ortho}{A character value (an English word or name)}
}
\value{
tot_syls The number of syllables in the string
}
\description{
How many syllables are in an English word?
}
\examples{
english_syllable_count('Mississippi')
}
\seealso{
Full credit for the development of this function goes to Tyler Kendall at the University of Oregon http://lingtools.uoregon.edu/scripts/english_syllable_counter-102.R
}
\keyword{text}
