% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_emoticon.R
\name{find_emoticon}
\alias{find_emoticon}
\title{Emoticon count}
\usage{
find_emoticon(x)
}
\arguments{
\item{x}{A string value}
}
\value{
vals The count of emoticons in a string
}
\description{
Count the number of emoticons in a string value
}
\examples{
find_emoticon("I added some emoticons =] :D")
}
\seealso{
http://cool-smileys.com/text-emoticons
}
\keyword{text}
