% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_method_2_function.R
\name{gender_method_2}
\alias{gender_method_2}
\title{Predict gender with word and character unigrams}
\usage{
gender_method_2(input, labeled = FALSE, set.seed = NULL,
  training_set = 0.8, chunk = FALSE, action = "append", dir = NULL)
}
\arguments{
\item{input}{A data frame containing Twitter profile information. Created using the function get_user_info.}

\item{labeled}{Set to TRUE if you have labeled data and want to train/test the algorithm on your data. Defautls to FALSE.}

\item{set.seed}{If labeled==TRUE, enter a seed to replicate creation of test/training data. Defaults to NULL.}

\item{training_set}{If labeled==TRUE, what proportion of your data will be your training set?}

\item{chunk}{If TRUE, features and predictions are generated for 10,000 user chunks. Recommended for unlabeled data with >10,000 users.}

\item{action}{If chunk==TRUE, whether the data should be appended and returned as one file, or saved as chunks. Options are append and save.}

\item{dir}{Where should output files be saved? Used only if action==save. Defaults to NULL}
}
\value{
results A two-column data frame containing Twitter user IDs and gender predictions.
}
\description{
This function inputs a data frame containing user metadata, builds features containing word and character unigrams from users' names, and employs an SVM classifier to predict gender.
}
\examples{
## df_list from function get_user_info
## df <- do.call("rbind", df_list)
gender_method_2(df, labeled=FALSE, set.seed=NULL, training_set=0.80, chunk=TRUE, action="save", dir="/home/User/workingDir")
}
\seealso{
Prediction method is based on: Burger, J.D. et al. 2011. Discriminating gender on Twitter. Proceedings of the Conference on Empirical Methods in Natural Language Processing (2011), 1301-1309.
}
\keyword{demography}
\keyword{gender,}
\keyword{twitter,}
