% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_method_3_function.R
\name{gender_method_3}
\alias{gender_method_3}
\title{Predict gender using structural name features}
\usage{
gender_method_3(input, labeled = FALSE, set.seed = NULL,
  training_set = 0.8, chunk = FALSE, action = "append", dir = NULL)
}
\arguments{
\item{input}{A data frame containing Twitter profile information. Created using the function get_user_info.}

\item{labeled}{Set to TRUE if you have labeled data and want to train/test the algorithm on your data. Defautls to FALSE.}

\item{set.seed}{If labeled==TRUE, enter a seed to replicate creation of test/training data. Defaults to NULL.}

\item{training_set}{If labeled==TRUE, what proportion of your data will be your training set?}

\item{chunk}{If TRUE, features and predictions are generated for 10,000 user chunks. Recommended for unlabeled data with >10,000 users.}

\item{action}{If chunk==TRUE, whether the data should be appended and returned as one file, or saved as chunks. Options are append and save.}

\item{dir}{Where should output files be saved? Used only if action==save. Defaults to NULL}
}
\value{
A two-column data frame containing Twitter user IDs and gender predictions.
}
\description{
This function inputs a data frame containing user metadata, builds features based on the linguistic structure of the users' names, and employs a decision tree classifier to predict gender.
}
\examples{
## df_list from function get_user_info
## df <- do.call("rbind", df_list)
#' gender_method_3(df, labeled=FALSE, set.seed=NULL, training_set=0.80, chunk=TRUE, action="save", dir="/home/User/workingDir")
}
\seealso{
Prediction method is based on: Mueller, J. and Stumme, G. 2016. Gender Inference using Statistical Name Characteristics in Twitter. Proceedings of the The 3rd Multidisciplinary International Social Networks Conference on SocialInformatics, Albany, NY, 2016, 1-8.
}
\keyword{demography}
\keyword{gender,}
\keyword{twitter,}
