% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitCred.R
\name{get_twitCred}
\alias{get_twitCred}
\title{Authenticate connection to the Twitter API: Part I}
\usage{
get_twitCred(api_key, api_secret)
}
\arguments{
\item{api_key}{The Consumer Key value associated with your Twitter app}

\item{api_secret}{Consumer secret associated with your Twitter app}
}
\value{
twitCred An authentication token
}
\description{
This function allows users to authenticate a connection to the Twitter API. To establish an API connection, you will need a Twitter Developer's Account and an application. See https://apps.twitter.com/ for details.
}
\keyword{twitter}
