% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_info.R
\name{get_user_info}
\alias{get_user_info}
\title{Predict gender using name word and character n-grams}
\usage{
get_user_info(user_ids, credential = NULL, rawdata = FALSE,
  datadir = "", is_id = "")
}
\arguments{
\item{user_ids}{A vector of Twitter user ID values or screen names}

\item{credential}{Your API authentication token}

\item{rawdata}{Whether you want to store the raw .json output}

\item{datadir}{Where you would like hte raw .json output to be stored}

\item{is_id}{TRUE if your input is ID values. FALSE if your input is screen names}
}
\value{
uInfo A data frame containing detailed profile metadata
}
\description{
This function does the following  1.) Take input data (as a data frame) 2.) Train a classifier using Kaggle data  3.) Generates predictions from the input data and classifier
}
\examples{
get_user_info(df$id_str, credential=user.signature, rawdata=FALSE, datadir="", is_id=TRUE)
}
\keyword{twitter}
