% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUpperMethod1.R
\name{isUpperMethod1}
\alias{isUpperMethod1}
\title{Detect upper case letters}
\usage{
isUpperMethod1(s)
}
\arguments{
\item{vect}{A single letter}
}
\value{
result A TRUE/FALSE value designating whether the letter is upper case
}
\description{
Is the letter upper case?
}
\examples{
isUpperMethod1("H")
}
\seealso{
This method originally proposed by StackOverflow user topchef
}
\keyword{text}
