% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isVowel.R
\name{isVowel}
\alias{isVowel}
\title{Is it a vowel?}
\usage{
isVowel(char)
}
\arguments{
\item{char}{A single letter}
}
\value{
A TRUE/FALSE value indicating whether the letter is a vowel
}
\description{
Evaluates whether a letter is a vowel
}
\examples{
isVowel("d")
}
\keyword{language}
\keyword{text,}
