% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_lda_topic.R
\name{lda_topic_4_user}
\alias{lda_topic_4_user}
\title{Assign topics to users}
\usage{
lda_topic_4_user(dat, topicnum, feature, identifier)
}
\arguments{
\item{dat}{A data frame containing Twitter profile information. Created using the function get_user_info.}

\item{topicnum}{The number of LDA topics to identify}

\item{feature}{Select text feature to analyze. Options are "laststatus_text" and "description."}

\item{identifier}{Select feature for identifying users. Options are "id_str" and "screen_name."}
}
\value{
ldaout_topic A two-column data frame containing either screen names or ID values of Twitter users linked to an assigned topic.
}
\description{
Identify topic within Twitter profile text fields, and assign topics to individual users.
}
\seealso{
: The LDA analysis framework employed in this function is based on instructions from Kailash Awati of Sensanalytics. See: https://eight2late.wordpress.com/2015/09/29/a-gentle-introduction-to-topic-modeling-using-r/
}
\keyword{Twitter}
\keyword{text,}
