% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{A vector of string or numeric values}
}
\value{
The most frequently occuring value
}
\description{
Find modal value, excluding NA values
}
\examples{
Mode(c(NA, 3, 4, 3, 5, NA))
}
\seealso{
Originally posted on StackOverflow by user Ken Williams on November 18, 2011
}
\keyword{statistics}
\keyword{summary}
