% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostCom.R
\name{mostCom}
\alias{mostCom}
\title{Find most frequently occuring value}
\usage{
mostCom(vect)
}
\arguments{
\item{vect}{A vector of string or numeric values}
}
\value{
result The most frequently occuring value, omitting NA values
}
\description{
Which value occurs most often, omitting NA values?
}
\examples{
mostCom(c(NA, 3, 4, 3, 5, NA))
}
\seealso{
Function author: Gilad Amitai
}
\keyword{statistics}
\keyword{summary}
