% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateFromLower.R
\name{stateFromLower}
\alias{stateFromLower}
\title{Lengthen state abbreviations}
\usage{
stateFromLower(x)
}
\arguments{
\item{x}{A string representing a state name abbreviation (i.e. "AK")}
}
\value{
refac.x A full state name
}
\description{
Convert abbreviated state names to full state names. Retrieved from:  Retrieved from: https://favorableoutcomes.wordpress.com/tag/state/
}
\examples{
stateFromLower("AK")
}
\seealso{
Code written by Christopher O'Brien, retrieved from: https://favorableoutcomes.wordpress.com/tag/state/
}
\keyword{geolocation}
