% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateToLower.R
\name{stateToLower}
\alias{stateToLower}
\title{Abbreviate state names}
\usage{
stateToLower(x)
}
\arguments{
\item{X}{A string representing a state name (i.e. "Alaska")}
}
\value{
refac.x Abbreviated state name
}
\description{
Convert full state names to abbreviations
}
\examples{
stateToLower("Alaska")
}
\seealso{
Code written by Christopher O'Brien, retrieved from: https://favorableoutcomes.wordpress.com/tag/state/
}
\keyword{geolocation,}
\keyword{state}
