% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim trailing and leading whitespace from a string}
\usage{
trim(x)
}
\arguments{
\item{x}{a string value}
}
\value{
A trimmed sentence
}
\description{
Trim trailing and leading whitespace from a string
}
\examples{
trim(" Trim this sentence ")
}
\keyword{text}
