% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vowel_count.R
\name{vowel_count}
\alias{vowel_count}
\title{Count vowels in a string}
\usage{
vowel_count(name)
}
\arguments{
\item{name}{A string value}
}
\value{
vowel_count A numeric value representing the number of vowels
}
\description{
Count the number of vowels in a string
}
\examples{
vowel_count("Jim")
}
\keyword{language}
\keyword{text,}
